+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/CHANGELOG.md
@@ -7,6 +7,10 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+## 1.17.0-beta.1
+
+Released 2026-Jul-16
+
 * Added a verbose-level diagnostic event for ignored metrics.
   ([#7429](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7429))
 
@@ -65,6 +69,16 @@ Notes](../../RELEASENOTES.md).
   when an `application/openmetrics-text` `Accept` header entry does not
   specify a `version` parameter.
   ([#7491](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7491))
+
+* Added `PrometheusHttpListenerOptions.TranslationStrategy` to control how
+  OpenTelemetry metric and label names are translated into Prometheus names.
+  ([#7507](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7507))
+
+* Fix stack overflow during metric collection when under load.
+  ([#7524](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7524))
+
+* Updated OpenTelemetry core component version(s) to `1.17.0`.
+  ([#7530](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7530))
 
 ## 1.16.0-beta.1